Ajout d'un test sur getPrimeAnnuelle des Managers
